refactor(node): flatten genesis package and tidy config loader
The genesis package held a single module, so flatten it to node/genesis.py
beside its consumer node/anchor.py. The chain package stays focused on
consensus timing.

Tidy the loader itself:
- load the file with Path.read_text instead of an explicit handle
- derive the hex pad width from the key length instead of a magic literal
- bound the validator set before decoding keys (mode="before"), matching
  the memory-safety rationale the check claims
- trim the documentation to the leanest form the rules allow
