--- src/lean_spec/node/genesis/__init__.py
@@ -1,5 +1,0 @@
-"""Genesis configuration and state initialization."""
-
-from lean_spec.node.genesis.config import GenesisConfig
-
-__all__ = ["GenesisConfig"]
