--- tests/node/genesis/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the genesis configuration module."""
